#6: Se separan los archivos de views en archivos independientes para cada pagina
