+++ MUSSA_Flask/app/templates/base.html
@@ -9,6 +9,7 @@
         | <a href="{{ url_for('logout') }}">Logout</a>
         {% endif %}
     </div>
+    -->
     <hr>
     {% with messages = get_flashed_messages() %}
     {% if messages %}
@@ -19,7 +20,7 @@
     </ul>
     {% endif %}
     {% endwith %}
--->
+    
     {% block content %}{% endblock %}
   </body>
 </html>
