+++ MUSSA_Flask/app/templates/login.html
@@ -1,10 +1,10 @@
 <link rel="stylesheet" href="/static/login_style.css" type="text/css">
 {% block body %}
+
 {% if session['logged_in'] %}
 <p>You're logged in already!</p>
 {% else %}
- 
- 
+  
 <form action="/login" method="POST">
     <div class="login">
         <div class="login-screen">
@@ -31,4 +31,5 @@
 </form>
  
 {% endif %}
+
 {% endblock %}
